requirements.txt
pandas
numpy
scikit-learn
matplotlib
streamlit
psycopg2-binary
python-dotenv
seaborn
